requirements.txt
mysql-connector-python==9.3.0
python-dotenv==1.0.1
ttkbootstrap==1.10.1
bcrypt==4.0.1